Increased stability of vertex fits (#16)
* Fixed stability of CHI2 estimator in vertex fits
* Changed default estimator for vertex fits

The many outliers far from the gaussian peak disturb the chi2 estimator
making the fit instable and crash. Which ultimately give wrong results.
This is stabalized further by using an automatic fit range chosen closer
around the peak region, effectively cutting away outliers.

It should be noted that this is not a 100% guarantee for stability and
the LOG_LIKELIHOOD estimator should be chosen instead.
Therefore, the default estimator for vertex fits is now LogLikelihood.