feat: add option to skip IP determination
Extend the determineLuminosity script to be able to skip IP
determination. By default a scenario which is misaligned and not
aligned will skip IP determination and an IP at the origin is assumed
instead. Alternatively the argument `--disable_ip_determination`
provides a user handle to force this behavior on all processed
scenarios.