refactor: handle arbitrary data sources in data reading module  (#21)
* dev: add vscode c++ configuration settings

* refactor!: remove PndTrackQ from data interface

The DataReader abstract class used the PndTrackQ in the interface an as
a internal structure, which does not allow for for compilations without
Pandaroot. Replace this data structure with other data structures
defined in data/TrackData.h.

Note: Separates the Pandaroot data reader into a own library.

* refactor!: use DataReader interface in DataFacade

Use the abstract DataReader class in the DataFacade UI. This enables
users to define custom DataReaders and inject them into the facade.